Update model, controller and urls for handling orders
