+++ server/src/models/order.ts
@@ -1,29 +1,6 @@
-/* eslint-disable camelcase */
-import mongoose from 'mongoose';
-import { CartSchema, ICart } from '../schemas/user';
+import CreateModel from './model';
+import OrderSchema, { IOrder } from '../schemas/order';
 
-export interface IOrder extends mongoose.Document {
-    _id: mongoose.Types.ObjectId;
-    name: string;
-    user_id: string;
-    email: string;
-    address: string;
-    postcode: string;
-    city: string;
-    country: string;
-    cart: ICart[];
-}
+const Order = new CreateModel<IOrder>('orders', OrderSchema).create();
 
-export const Order = mongoose.model<IOrder>(
-    'orders',
-    new mongoose.Schema({
-        name: String,
-        user_id: String,
-        email: String,
-        address: String,
-        postcode: Number,
-        city: String,
-        country: String,
-        cart: { type: [CartSchema] },
-    }),
-);
+export default Order;
