+++ server/src/routes/order.routes.ts
@@ -1,11 +1,27 @@
 import { Router } from 'express';
 import passport from 'passport';
 
-import { createOrderForUser, getOrdersByUserId } from '../controllers/order.controller';
+import { create, getAll } from '../controllers/common';
+import { IOrder } from '../schemas/order';
+import Order from '../models/order';
+import { IUser } from '../schemas/user';
 
 const router = Router();
 
-router.post('/', passport.authenticate('jwt', { session: false }), createOrderForUser);
-router.get('/:id/get-orders', passport.authenticate('jwt', { session: false }), getOrdersByUserId);
+router.post('/', create<IOrder>(Order));
+router.get(
+    '/',
+    (req, res, next) => {
+        passport.authenticate('jwt', { session: false }, (err, user) => {
+            if (user) {
+                req.query.user_id = (user as IUser)?._id.toString();
+                next();
+            } else {
+                res.status(401).send('Unauthorized');
+            }
+        })(req, res, next);
+    },
+    getAll<IOrder>(Order),
+);
 
 export default router;
